+++ example/lib/call_screen.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
+import 'dart:io';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:twilio_voice/twilio_voice.dart';
 
@@ -12,6 +14,7 @@ class _CallScreenState extends State<CallScreen> {
   var speaker = false;
   var mute = false;
   var isHolding = false;
+  var isBluetoothOn = false;
   var isEnded = false;
 
   String message = "";
@@ -119,6 +122,7 @@ class _CallScreenState extends State<CallScreen> {
       _updateMuteState(),
       _updateSpeakerState(),
       _updateHoldState(),
+      _updateBluetoothState(),
     ]);
   }
 
@@ -141,6 +145,13 @@ class _CallScreenState extends State<CallScreen> {
     final isHolding = await TwilioVoice.instance.call.isHolding();
     setState(() {
       this.isHolding = isHolding ?? false;
+    });
+  }
+
+  Future<void> _updateBluetoothState() async {
+    final isBluetoothOn = await TwilioVoice.instance.call.isBluetoothOn();
+    setState(() {
+      this.isBluetoothOn = isBluetoothOn ?? false;
     });
   }
 
@@ -219,6 +230,47 @@ class _CallScreenState extends State<CallScreen> {
                                     .toggleSpeaker(!speaker).then((value) {
                                   _updateSpeakerState();
                                 });
+                              },
+                            ),
+                          ),
+                        ),
+                        Material(
+                          type: MaterialType
+                              .transparency, //Makes it usable on any background color, thanks @IanSmith
+                          child: Ink(
+                            decoration: BoxDecoration(
+                              border:
+                                  Border.all(color: Colors.white, width: 1.0),
+                              color: mute
+                                  ? Theme.of(context).colorScheme.secondary
+                                  : Colors.white24,
+                              shape: BoxShape.circle,
+                            ),
+                            child: InkWell(
+                              //This keeps the splash effect within the circle
+                              borderRadius: BorderRadius.circular(
+                                  1000.0), //Something large to ensure a circle
+                              child: Padding(
+                                padding: EdgeInsets.all(20.0),
+                                child: Icon(
+                                  Icons.bluetooth,
+                                  size: 40.0,
+                                  color: Colors.white,
+                                ),
+                              ),
+                              onTap: () {
+                                // BLuetooth functionality is not supported on web or macos
+                                if(kIsWeb || Platform.isMacOS) {
+                                  return;
+                                }
+                                final newState = !isBluetoothOn;
+                                print("bluetooth? ${newState ? "on" : "off"}");
+                                TwilioVoice.instance.call.toggleBluetooth(bluetoothOn: newState).then((value) {
+                                  _updateBluetoothState();
+                                });
+                                // setState(() {
+                                //   mute = !mute;
+                                // });
                               },
                             ),
                           ),