+++ lib/_internal/method_channel/twilio_call_method_channel.dart
@@ -100,4 +100,14 @@ class MethodChannelTwilioCall extends TwilioCallPlatform {
   Future<bool?> sendDigits(String digits) {
     return _channel.invokeMethod('sendDigits', <String, dynamic>{"digits": digits});
   }
+
+  @override
+  Future<bool?> toggleBluetooth({bool bluetoothOn = true}) {
+    return _channel.invokeMethod('toggleBluetooth', <String, dynamic>{"bluetoothOn": bluetoothOn});
+  }
+
+  @override
+  Future<bool?> isBluetoothOn() {
+    return _channel.invokeMethod('isBluetoothOn', <String, dynamic>{});
+  }
 }