+++ lib/_internal/twilio_voice_web.dart
@@ -363,6 +363,20 @@ class TwilioVoiceWeb extends MethodChannelTwilioVoice {
     }
   }
 
+  /// Request bluetooth permissions.
+  /// Not supported on web.
+  @override
+  Future<bool?> requestBluetoothPermissions() async {
+    return true;
+  }
+
+  /// Queries browser for bluetooth permissions.
+  /// Not supported on web.
+  @override
+  Future<bool> hasBluetoothPermissions() async {
+    return true;
+  }
+
   /// Request notifications permission. Returns true if permission is granted, false otherwise.
   /// Documentation: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/permissions/request
   @override
@@ -610,6 +624,18 @@ class Call extends MethodChannelTwilioCall {
       return true;
     }
     return false;
+  }
+
+  /// Not available for web
+  @override
+  Future<bool> toggleBluetooth({bool bluetoothOn = true}) {
+    return Future.value(bluetoothOn);
+  }
+
+  /// Not available for web
+  @override
+  Future<bool> isBluetoothOn() {
+    return Future.value(false);
   }
 
   /// Not currently implemented for web